+++ zerver/actions/create_user.py
@@ -306,7 +306,7 @@ def notify_created_user(user_profile: UserProfile) -> None:
     user_ids_without_real_email_access = []
     for user in active_users:
         if can_access_delivery_email(
-            user, user_profile.id, user_profile.realm.email_address_visibility
+            user, user_profile.id, user_profile.realm.email_address_visibility, user_row["is_bot"]
         ):
             user_ids_with_real_email_access.append(user.id)
         else:
@@ -322,7 +322,7 @@ def notify_created_user(user_profile: UserProfile) -> None:
         )
 
     if user_ids_without_real_email_access:
-        del person["delivery_email"]
+        person["delivery_email"] = None
         event = dict(type="realm_user", op="add", person=person)
         transaction.on_commit(
             lambda event=event: send_event(